feat: add raw streams and rpcstream proxy
Add a mechanism for opening a raw stream without initializing the call.

Use io.ReadWriteCloser in more locations for more flexibility.

Add a proxy to rpcstream for proxying between two RpcStream services.

Instead of requiring the stream implementation to parse to a Packet struct, pass
the raw data.

This is helpful in situations like rpcstream where we want to proxy the data
through without slowing things down by parsing the packets unnecessarily.

Use raw message encoding in the rpcstream to improve performance.

Re-use the memory buffer when proxying rpcstreams.

Signed-off-by: Christian Stewart <christian@paral.in>